refactor: split browser sdk imports for sandbox and audit
