+++ extensions/browser/browser-bridge.ts
@@ -1,0 +1,2 @@
+export type { BrowserBridge } from "./src/browser/bridge-server.js";
+export { startBrowserBridgeServer, stopBrowserBridgeServer } from "./src/browser/bridge-server.js";
