+++ extensions/browser/browser-cdp.ts
@@ -1,0 +1,1 @@
+export { redactCdpUrl } from "./src/browser/cdp.helpers.js";
