+++ extensions/browser/browser-config.ts
@@ -1,0 +1,7 @@
+export {
+  DEFAULT_BROWSER_EVALUATE_ENABLED,
+  DEFAULT_OPENCLAW_BROWSER_COLOR,
+  DEFAULT_OPENCLAW_BROWSER_PROFILE_NAME,
+} from "./src/browser/constants.js";
+export type { ResolvedBrowserConfig, ResolvedBrowserProfile } from "./src/browser/config.js";
+export { resolveBrowserConfig, resolveProfile } from "./src/browser/config.js";
