+++ extensions/browser/browser-control-auth.ts
@@ -1,0 +1,2 @@
+export type { BrowserControlAuth } from "./src/browser/control-auth.js";
+export { ensureBrowserControlAuth, resolveBrowserControlAuth } from "./src/browser/control-auth.js";
