+++ src/agents/sandbox/context.ts
@@ -3,9 +3,10 @@ import type { OpenClawConfig } from "../../config/config.js";
 import { loadConfig } from "../../config/config.js";
 import {
   DEFAULT_BROWSER_EVALUATE_ENABLED,
+  ensureBrowserControlAuth,
   resolveBrowserControlAuth,
-} from "../../plugin-sdk/browser-config.js";
-import { ensureBrowserControlAuth } from "../../plugin-sdk/browser-runtime.js";
+} from "../../plugin-sdk/browser-control-auth.js";
+import { DEFAULT_BROWSER_EVALUATE_ENABLED } from "../../plugin-sdk/browser-config.js";
 import { defaultRuntime } from "../../runtime.js";
 import { resolveUserPath } from "../../utils.js";
 import { syncSkillsToWorkspace } from "../skills.js";
