+++ src/agents/sandbox/manage.test.ts
@@ -19,9 +19,13 @@ const backendMocks = vi.hoisted(() => ({
   removeRuntime: vi.fn(),
 }));
 
-vi.mock("../../config/config.js", () => ({
-  loadConfig: configMocks.loadConfig,
-}));
+vi.mock("../../config/config.js", async (importOriginal) => {
+  const actual = await importOriginal<typeof import("../../config/config.js")>();
+  return {
+    ...actual,
+    loadConfig: configMocks.loadConfig,
+  };
+});
 
 vi.mock("./registry.js", () => ({
   readBrowserRegistry: registryMocks.readBrowserRegistry,
