+++ src/plugin-sdk/browser-bridge.ts
@@ -1,0 +1,5 @@
+export type { BrowserBridge } from "../../extensions/browser/browser-bridge.js";
+export {
+  startBrowserBridgeServer,
+  stopBrowserBridgeServer,
+} from "../../extensions/browser/browser-bridge.js";
