+++ src/plugin-sdk/browser-cdp.ts
@@ -1,0 +1,1 @@
+export { redactCdpUrl } from "../../extensions/browser/browser-cdp.js";
