+++ src/plugin-sdk/browser-control-auth.ts
@@ -1,0 +1,5 @@
+export type { BrowserControlAuth } from "../../extensions/browser/browser-control-auth.js";
+export {
+  ensureBrowserControlAuth,
+  resolveBrowserControlAuth,
+} from "../../extensions/browser/browser-control-auth.js";
